Add files via upload
Develop programs to implement the following system. The system consists of a foreground process and a background process. Both share an array of integers. 
The foreground process accepts integers through the terminal and appends each integer to the shared array.
The background process keeps the shared array sorted in ascending order after each addition. 
When the last element is also placed in its correct position, the foreground process displays the shared array on the terminal.